requirements.txt
# LangChain Core
langchain>=0.3.14
langchain-core>=0.3.28
langchain-community>=0.3.14
langchain-text-splitters>=0.3.4

# Google Gemini
langchain-google-genai>=2.0.8

# Vector Store
langchain-chroma>=0.2.2
chromadb>=0.5.23

# LangGraph
langgraph>=0.2.60
langgraph-checkpoint>=2.0.10

# Streamlit UI
streamlit>=1.41.0

# Document Loaders
pypdf>=5.1.0
unstructured>=0.16.11
markdown>=3.7

# Web Search (optional)
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.0
